created a component popup
